Created Menu class. Rrmoved all console functionality here. Remove decimal type usages

+++ ParkingEmulatorTask/Car.cs
@@ -5,10 +5,10 @@ namespace ParkingEmulatorTask
     class Car
     {
         public int Id { get; set; }
-        public decimal Balance { get; set; }
+        public double Balance { get; set; }
         public CarType CarType { get; set; }
 
-        public Car(decimal firstPayment , CarType carType)
+        public Car(double firstPayment , CarType carType)
         {
             Id = GenerateId();
             Balance = firstPayment;
